requirements.txt
python-telegram-bot
matplotlib
sqlite3
